Refactored project into edu.ensign.cs460
