--- src/main/java/com/example/explorecalijpa/model/Difficulty.java
@@ -1,14 +1,0 @@
-package com.example.explorecalijpa.model;
-
-/**
- * Created by maryellenbowman on 7/3/16.
- */
-
-/**
- * Enumeration of the level of effort.
- *
- * Created by Mary Ellen Bowman
- */
-public enum Difficulty {
-    Easy, Medium, Difficult, Varies;
-}
