--- src/main/java/com/example/explorecalijpa/model/Region.java
@@ -1,27 +1,0 @@
-package com.example.explorecalijpa.model;
-
-/**
- * Enumeration of the region of California.
- *
- * Created by Mary Ellen Bowman
- */
-public enum Region {
-    Central_Coast("Central Coast"), Southern_California("Southern California"),
-    Northern_California("Northern California"), Varies("Varies");
-    private String label;
-    private Region(String label) {
-        this.label = label;
-    }
-
-    public static Region findByLabel(String byLabel) {
-        for(Region r:Region.values()) {
-            if (r.label.equalsIgnoreCase(byLabel))
-                return r;
-        }
-        return null;
-    }
-
-    public String getLabel() {
-        return label;
-    }
-}
