+++ src/main/java/edu/ensign/cs460/model/Difficulty.java
@@ -1,0 +1,14 @@
+package edu.ensign.cs460.model;
+
+/**
+ * Created by maryellenbowman on 7/3/16.
+ */
+
+/**
+ * Enumeration of the level of effort.
+ *
+ * Created by Mary Ellen Bowman
+ */
+public enum Difficulty {
+    Easy, Medium, Difficult, Varies;
+}
