+++ src/main/java/edu/ensign/cs460/model/Tour.java
@@ -1,0 +1,151 @@
+package edu.ensign.cs460.model;
+
+import jakarta.persistence.*;
+
+import java.util.Objects;
+
+/**
+ * The Tour contains all attributes of an Explore California Tour.
+ *
+ * Created by Mary Ellen Bowman
+ */
+@Entity
+public class Tour {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column
+    private String title;
+
+    @Column(length = 2000)
+    private String description;
+
+    @Column(length = 2000)
+    private String blurb;
+
+    @Column
+    private Integer price;
+
+    @Column
+    private String duration;
+
+    @Column(length = 2000)
+    private String bullets;
+
+    @Column
+    private String keywords;
+
+
+    @ManyToOne
+    @JoinColumn(name="tour_package_code")
+    private TourPackage tourPackage;
+
+    @Column
+    @Enumerated(EnumType.STRING)
+    private Difficulty difficulty;
+
+    @Column
+    private Region region;
+
+    public Tour(String title, String description, String blurb, Integer price, String duration, String bullets,
+                String keywords, TourPackage tourPackage, Difficulty difficulty, Region region) {
+        this.title = title;
+        this.description = description;
+        this.blurb = blurb;
+        this.price = price;
+        this.duration = duration;
+        this.bullets = bullets;
+        this.keywords = keywords;
+        this.tourPackage = tourPackage;
+        this.difficulty = difficulty;
+        this.region = region;
+    }
+
+    protected Tour() {
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public String getBlurb() {
+        return blurb;
+    }
+
+    public Integer getPrice() {
+        return price;
+    }
+
+    public String getDuration() {
+        return duration;
+    }
+
+    public String getBullets() {
+        return bullets;
+    }
+
+    public String getKeywords() {
+        return keywords;
+    }
+
+    public TourPackage getTourPackage() {
+        return tourPackage;
+    }
+
+    public Difficulty getDifficulty() {
+        return difficulty;
+    }
+
+    public Region getRegion() {
+        return region;
+    }
+
+    @Override
+    public String toString() {
+        return "Tour{" +
+                "id=" + id +
+                ", title='" + title + '\'' +
+                ", description='" + description + '\'' +
+                ", blurb='" + blurb + '\'' +
+                ", price=" + price +
+                ", duration='" + duration + '\'' +
+                ", bullets='" + bullets + '\'' +
+                ", keywords='" + keywords + '\'' +
+                ", tourPackage=" + tourPackage +
+                ", difficulty=" + difficulty +
+                ", region=" + region +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Tour tour = (Tour) o;
+        return Objects.equals(id, tour.id) &&
+                Objects.equals(title, tour.title) &&
+                Objects.equals(description, tour.description) &&
+                Objects.equals(blurb, tour.blurb) &&
+                Objects.equals(price, tour.price) &&
+                Objects.equals(duration, tour.duration) &&
+                Objects.equals(bullets, tour.bullets) &&
+                Objects.equals(keywords, tour.keywords) &&
+                Objects.equals(tourPackage, tour.tourPackage) &&
+                difficulty == tour.difficulty &&
+                region == tour.region;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, title, description, blurb, price, duration, bullets, keywords, tourPackage, difficulty, region);
+    }
+}
