+++ src/main/java/edu/ensign/cs460/model/TourPackage.java
@@ -1,0 +1,60 @@
+package edu.ensign.cs460.model;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+import java.util.Objects;
+
+/**
+ * A Classification of Tours.
+ *
+ * Created by Mary Ellen Bowman
+ */
+@Table(name="tour_package")
+@Entity
+public class TourPackage {
+    @Id
+    private String code;
+
+    @Column
+    private String name;
+
+    protected TourPackage() {
+    }
+
+    public TourPackage(String code, String name) {
+        this.code = code;
+        this.name = name;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public String toString() {
+        return "TourPackage{" +
+                "code='" + code + '\'' +
+                ", name='" + name + '\'' +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        TourPackage that = (TourPackage) o;
+        return Objects.equals(code, that.code) &&
+                Objects.equals(name, that.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(code, name);
+    }
+}
