+++ src/main/java/edu/ensign/cs460/recommendation/TourSummary.java
@@ -1,0 +1,13 @@
+package edu.ensign.cs460.recommendation;
+
+/**
+ * Tour Summary interface
+ *
+ * Created by Antonio Martinez
+ */
+public interface TourSummary {
+    Integer getTourId();
+    String getTitle();
+    Double getAvgScore();
+    Long getReviewCount();
+}
