+++ src/main/resources/application.properties
@@ -8,3 +8,6 @@ spring.jpa.hibernate.ddl-auto=none
 
 # Disable docker compose
 spring.docker.compose.enabled=false
+
+#Set the server port to 8000
+server.port=8000
